arch: arm: Reorganise ARM interrupt management routines.
This commit reorganises the existing arch/arm/irq_manage.c file into
three separate files:

* arch/arm/irq.c: Common IRQ handlers and management routines
* arch/arm/cortex_m/irq_manage.c: Cortex-M IRQ management routines
* arch/arm/cortex_r/irq_manage.c: Cortex-R IRQ management routines
